requirements.txt
python-telegram-bot >= 20.2
Pillow >= 9.4.0
